feat: add loading of battle data, improve map load and render
This commit adds functionality for loading an Ars Militaris battle from a .csv
file, and also improves the functionality for loading and rendering a map.

The Ars Militaris scenario currently being developed is The Patrol: Ambush. Its
battle data is in the 'the_patrol_ambush_data.csv' file. Its battle data is
loaded and Bevy ECS Entities are created for the units.

For the map, a change has been made to instead of simply rendering the read
file, the system setup_map_system now creates an Entity for each tile, with a
2D UI Text component and a Tile component.

Some older systems have been cleaned up.