feat: add loading of unit IDs to map
This commit adds a new system, place_units_on_map_system, that queries the
in-memory representation of the map and the battle data, and places each unit
(in truth, its Unit ID) on the displayed map.

The map.txt file was changed to eliminate the existing Unit ID there as it was
no longer relevant.

The the_patrol_ambush_data.csv was modified because the previous version had the
Gaul units out of bounds.

There is currently a "problem" with the implementation, in that the
place_units_on_map_system system should be run at startup, but when declared to
do so, there's no guarantees that it will run after the map is loaded in the
setup_map_system system.