feat: improve game control logic
This commit improves the project with the addition of Bevy Events and States to
better control the execution of systems.

The game now waits in a "Main Menu" for the player to press the Space button,
and then loads the battle and displays it on the screen.

Two times has happened that the units weren't loaded. This is strange because it
is inconsistent. It will be important to keep awareness of this behavior.