feat: add cursor movement system
A new system, move_cursor_system, is added to handle the movement of the cursor
on the map. Functionality has already been implemented to move the cursor in all
the four directons, with 'WASD'.

The implementation already detects the edge of the map and won't allow the
cursor to go past there.