feat: add wait turn mechanic
This commit adds the wait_turn_system and the end_turn_system systems that
implement the functionality of the Wait Turn mechanic in Ars Militaris.

The wait_turn_system system decreases all units WT until it finds an unit with
WT of 0. It then sets that unit as the current unit (in a new Game Resource),
and returns the game to the cursor stage (or rather, the Battle State).

The end_turn_system system waits for the player to press 'T' to end the current
unit's turn. But currently, there's a bug with changing between Bevy States.

We're now going to try the iyes_loopless crate's implementation to see if it
works.