feat: use iyes_loopless crate
This commit replaces the default Bevy SystemSet calls with iyes_loopless ones.

Some systems were updated to reflect the new usage of iyes_loopless States.

With this, the Wait Turn mechanic is now working correctly, except for one small
detail, in that, because the WT is not being reset for the units once they end
their turn, this creates unwanted behavior.