feat: use bevy_quinnet
This commit marks the successful integration of the 'bevy_quinnet' crate into
the project.

An amclient separate project has now been split from this one, and development
will occur in parallel.

Currently, the server listens to a StartGame message from the client, and, upon
receiving it, sends it back to the client. That's enough functionality for a
start.