feat: change wait turn behavior
This commit changes the behavior of the wait_turn_system and its synchronization
with the clients.

Now, the wait_turn_system only runs on the server, and, when complete, it will
send the updated WTs to the clients, along with the PlayerTurn message, with
data stating what player turn it is.