feat: add `Move` `UnitAction` synchronization
This commit adds the ability for the server to correctly process
`Move` `UnitAction`s sent by the players.

Now, when a player issues a `Move` `UnitAction`, it will be sent
to the server, and the server will broacast that `UnitAction` to
all the other clients. And, when it is processed, both (or more)
clients will see the unit moving at the same time.

Now, the only thing missing to have the Multiplayer aspect of Ars
Militaris in a good state is to have the `BasicAttack`
`UnitAction` processed correctly by the server.